pyproject.toml
[tool.poetry]
name = "telegram-cloudpayments-polling"
version = "0.1.0"
description = ""
authors = ["BuildTools <unconfigured@null.spigotmc.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
aiogram = "2.25"
pydantic = "^2.9.2"
requests = "^2.32.3"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
